Converted apu_test file to C++, added register notes
